requirements.txt
pandas
opencv-contrib-python
pillow
matplotlib
numpy